Link popup to comment button
