+++ src/accessButton.js
@@ -1,0 +1,8 @@
+import displayPop from './comment.js';
+
+const accessComment = () => {
+  const popUp = document.querySelector('.pop');
+  popUp.addEventListener('click', displayPop);
+};
+
+export default accessComment;
