+++ src/index.html
@@ -20,7 +20,7 @@
      </div>
   </header>
   <main>
-    <button class="pop" type="submit">Comment</button>  
+    <!-- <button class="pop" type="submit">Comment</button> -->
     <ul class="meal-list">
     </ul>
   </main>
